pyproject.toml
[project]
name = "mcp-guide"
version = "0.5.0"
description = "MCP server for handling guidelines, project rules and controlled development"
readme = "README.md"
license = {file = "LICENSE.md"}
authors = [{name = "David Nugent", email = "davidn@uniquode.io"}]
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.16.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    "aioshutil>=1.6",
    "aiofiles>=23.2.1",
    "chevron>=0.14.0",
    "click>=8.3.0",
    "uuid7>=0.1.0",
]

[project.scripts]
mcp-guide = "mcp_guide.main:main"
osvcheck = "scripts.osvcheck:main"

[tool.setuptools.package-data]
"*" = ["templates/**/*"]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.uv]
package = true

[tool.uv.build-backend]
module-name = ["mcp_core", "mcp_guide"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
filterwarnings = ["error"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/mcp_core",
    "--cov=src/mcp_guide",
    "--cov-report=term-missing",
]
markers = [
    "e2e: marks tests as end-to-end tests",
    "slow: marks tests as slow running",
]

[tool.ruff]
src = ["src"]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["I", "F401"]  # I = isort, F401 = unused imports

[tool.ruff.lint.isort]
known-first-party = ["mcp_core", "mcp_guide"]

[tool.mypy]
python_version = "3.13"
mypy_path = "src"
files = ["src"]
strict = true
namespace_packages = true

[[tool.mypy.overrides]]
module = "fastmcp"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "uuid_extensions"
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "coverage>=7.10.7",
    "ruff>=0.13.2",
    "mypy>=1.18.2",
    "pre-commit>=4.0.0",
    "types-aiofiles>=24.1.0.20250822",
    "types-pyyaml>=6.0.12.20250915",
    "types-chevron>=0.14.2.20250103",
    "watchdog>=6.0.0",
    "bandit>=1.9.2",
    "ty>=0.0.2",
    "pyright>=1.1.407",
]
debug = [
    "debugpy>=1.8.0",
]
